QRadar default config, preview directive, default daemon

+++ lib/saq/qradar.py
@@ -7,6 +7,8 @@ import datetime
 import logging
 import threading
 import time
+import io
+import json
 
 import requests
 
@@ -21,6 +23,11 @@ STATUS_SORTING = 'SORTING'
 STATUS_COMPLETED = 'COMPLETED'
 STATUS_CANCELED = 'CANCELED'
 STATUS_ERROR = 'ERROR'
+
+def format_qradar_datetime(t):
+    """Format the given datatime object into a string suitable for AQL."""
+    assert isinstance(t, datetime.datetime)
+    return t.strftime('%Y-%m-%d %H:%M %z')
 
 class QueryTimeoutError(RuntimeError):
     pass
@@ -60,7 +67,13 @@ class QRadarAPIClient(object):
         # used to wait until the next attempt to query the status of a search
         self.wait_control_event = threading.Event()
 
-    def execute_aql_query(self, query, timeout=None, query_timeout=None, delete=True, status_callback=None, continue_check_callback=None, retries=5):
+    def execute_aql_query(self, query, 
+                                timeout=None, 
+                                query_timeout=None, 
+                                delete=True, 
+                                status_callback=None, 
+                                continue_check_callback=None, 
+                                retries=5):
         self.cancel_flag = False
         self.wait_control_event.clear()
 
@@ -87,14 +100,16 @@ class QRadarAPIClient(object):
             # was the query canceled?
             if self.cancel_flag:
                 logging.warning(f"query {search_id} canceled")
-                self.delete_aql_query(search_id)
+                if delete:
+                    self.delete_aql_query(search_id)
                 raise QueryCanceledError(f"query {search_id} canceled")
 
             # has the query been executing for too long?
             if self.query_limit is not None:
                 if datetime.datetime.now() > self.query_limit:
                     logging.error(f"query {query} search_id {search_id} timed out")
-                    self.delete_aql_query(search_id)
+                    if delete:
+                        self.delete_aql_query(search_id)
                     raise QueryTimeoutError()
 
             # should we keep checking?
@@ -132,7 +147,8 @@ class QRadarAPIClient(object):
             
             if response.status_code < 200 or response.status_code > 299:
                 logging.error(f"unexpected error code for query {search_id}: {response.status_code} {response.text}")
-                self.delete_aql_query(search_id)
+                if delete:
+                    self.delete_aql_query(search_id)
                 raise QueryError(f"{response.status_code}:{response.text}")
 
             self.status_json = response.json()
@@ -152,7 +168,8 @@ class QRadarAPIClient(object):
 
             # did it get cancelled or error out
             if self.status_json[KEY_STATUS] in [ STATUS_CANCELED, STATUS_ERROR ]:
-                self.delete_aql_query(search_id)
+                if delete:
+                    self.delete_aql_query(search_id)
                 raise QueryError(f"query status {self.status_json[KEY_STATUS]}")
 
             # otherwise it completed
@@ -161,10 +178,14 @@ class QRadarAPIClient(object):
 
             if response.status_code == 200:
                 self.result_json = response.json()
-                self.delete_aql_query(search_id)
+                if delete:
+                    self.delete_aql_query(search_id)
+
                 return self.result_json
 
-            self.delete_aql_query(search_id)
+            if delete:
+                self.delete_aql_query(search_id)
+
             raise QueryError(f"search result download returned {response.status_code}: {response.text}")
 
     def cancel_aql_query(self):
@@ -179,3 +200,44 @@ class QRadarAPIClient(object):
             logging.debug(f"got result {response.status_code} for deletion of {search_id}")
         except Exception as e:
             logging.error(f"unable to delete query {search_id}: {e}")
+
+    def get_siem_offenses(self, range=None, fields=None, sort=None, filter=None):
+        """Retrieve a list of offenses currently in the system."""
+        assert filter is None or (isinstance(filter, str) and filter)
+        params = {}
+        if filter is not None:
+            params['filter'] = filter
+
+        return requests.get(f'{self.url}/siem/offenses', params=params,
+            headers=self.headers,
+            verify=False).json()
+
+    def close_siem_offense(self, offense_id, closing_reason_id):
+        """Sets the status field of the given offense to CLOSED with the given closing_reason_id."""
+        assert isinstance(offense_id, int)
+        assert isinstance(closing_reason_id, int)
+        return requests.post(f'{self.url}/siem/offenses/{offense_id}', params={
+            'status': 'CLOSED',
+            'closing_reason_id': str(closing_reason_id), },
+            headers=self.headers,
+            verify=False).json()
+
+    def get_offense_closing_reasons(self, filter=None):
+        """Queries for the closing reason that matches the given filter."""
+        assert filter is None or isinstance(filter, str)
+        params = {}
+        if filter is not None:
+            params['filter'] = filter
+
+        return requests.get(f'{self.url}/siem/offense_closing_reasons', 
+            params=params,
+            headers=self.headers,
+            verify=False).json()
+
+    def create_offense_closing_reason(self, reason):
+        """Creates a new closing reason with the given reason text."""
+        assert isinstance(reason, str) and reason
+        return requests.post(f'{self.url}/siem/offense_closing_reasons', params={
+            'reason': reason, },
+            headers=self.headers,
+            verify=False).json()
